Fixing name of UrlScalar
The README shows that there is a scalar named URL, but the implementation of the scalar is using "Url" as the name.  This causes an error when trying to declare and use the scalar like so:

```graphql
scalar URL
```

I can try to work around it, but this seems like a potential bug?